Refactor utils tests
